pass `config.yaml` into scripts from `run_me.sh`
The files in `./scripts` should be as portable as possible.

Previously, each script obtained project-config stuff (file paths etc)
from a `conf/config.yaml` file, the position of which was hard-coded
into the script.

With that approach, if a script was to be used (without modification) in
another project, a config file in the same relative position must be
available.

To improve the portability of the scripts, the filepath of the config is
now passed as a command line argument to each relevant script by a
coordinating script (here `run_me.sh`).

Also:
- add option parsing (using {optparse}) to the scripts