rewrite dupree scripts: map over single repo, then reduce the repos
Earlier dupree-analysis script was ran once and analysed each of the
repositories. This has been rewritten so that iterating over the
repositories is controlled by the calling process, and the dupree
analysis script just analyses a single repo.

Add explicit args (for single-repo script):
- `--local_repo` is the path to the repository to be analysed
- `--package_name` is the name of the analysed package (a `package`
column is added to the benchmark output);
- `--output_dir` is the path to the directory where the dupree-data will
be stored

Add explicit args (for dupree-collapsing script):
- `--input_files <file>` a single file containing the paths for all
files that should be combined together
- `--output` file as argument

Also:
- rename: dupree scripts are now 04a (analysis) and 04b (collapse)
- remove results-dir creation from dupree analysis script, this should
be done by the coordinating script