[smk] add `Snakefile` to run `gitsum`-analysis
We assume a list of repositories, and their URLs have been identified
before this Snakefile is called. The position of this file must be
specified in the config file.

The package-names and repository-locations (local & remote) are imported
into the Snakefile and are used to define the inputs to various rules.

Aims of design:
- limit 'implicit knowledge'
    - (coordinator shouldn't need to guess what files will be produced
when a rule is ran; that is, output files should be passed as arguments)
- configuration details (filepaths, disregarded repositories etc) should
be defined in one place and as-much-as-possible, be user-modifiable

Rule to collapse gitsum files (`collapse_gitsum`):
- the set of gitsum.tsv files that are to be collapsed can be computed
within the snakemake run, so these files are passed as a
trailing-argument to the `rowbind_tsv.R` script (rather than temporarily
making a file containing these filepaths)

Rule to run single-repo gitsum analysis (`single_repo_gitsum`):
- Uses a directory as `input`